* core/authentication_api.php   (auth_get_current_user_id): store an return an int, not a string (in case     we end up with more === comparisons)
git-svn-id: http://mantisbt.svn.sourceforge.net/svnroot/mantisbt/trunk@2039 f5dc347c-c33d-0410-90a0-b07cc1902cb9

+++ core/authentication_api.php
@@ -6,7 +6,7 @@
 	# See the README and LICENSE files for details
 
 	# --------------------------------------------------------
-	# $Id: authentication_api.php,v 1.27 2003-02-20 16:48:22 jlatour Exp $
+	# $Id: authentication_api.php,v 1.28 2003-02-26 09:11:19 jfitzell Exp $
 	# --------------------------------------------------------
 
 	###########################################################################
@@ -337,7 +337,7 @@
 			return false;
 		}
 
-		$t_user_id = db_result( $result );
+		$t_user_id = (int)db_result( $result );
 
 		$g_cache_current_user_id = $t_user_id;
